requirements.txt
python-telegram-bot==13.7
numpy==1.16.5
pandas==0.25.1
matplotlib==3.1.1